Created new Partial for Pagination Control
