+++ src/js/Pages/Settings/Partials/PaginationControl/PaginationControl.scss
@@ -1,0 +1,7 @@
+.pagination-control-container {
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+
+  max-width: 30%;
+}
