+++ src/js/Pages/Settings/Settings.scss
@@ -20,12 +20,6 @@
     }
   }
 
-  .pagination-control-container {
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-  }
-
   .checkbox {
     display: flex;
     flex-direction: row;
